simplified the API, modularized the lib, added full path support
