+++ lib/functions/bezier.js
@@ -1,0 +1,872 @@
+var Intersection = require('./Intersection');
+
+var affine = require('kld-affine');
+var Point2D = affine.Point2D;
+var Vector2D = affine.Vector2D;
+
+var Polynomial = require('kld-polynomial').Polynomial;
+
+function removeMultipleRootsIn01(roots) {
+    var ZEROepsilon = 1e-15;
+    roots.sort(function (a, b) { return a - b; });
+    for (var i = 1; i < roots.length;) {
+        if (Math.abs(roots[i] - roots[i - 1]) < ZEROepsilon) {
+            roots.splice(i, 1);
+        }
+        else {
+            i++;
+        }
+    }
+}
+
+module.exports = {};
+
+
+/**
+ *  intersectBezier2Bezier2
+ *
+ *  @param {Point2D} a1
+ *  @param {Point2D} a2
+ *  @param {Point2D} a3
+ *  @param {Point2D} b1
+ *  @param {Point2D} b2
+ *  @param {Point2D} b3
+ *  @returns {Intersection}
+ */
+module.exports.intersectBezier2Bezier2 = function(a1, a2, a3, b1, b2, b3) {
+    var a, b;
+    var c12, c11, c10;
+    var c22, c21, c20;
+    var result = new Intersection();
+    var poly;
+
+    a = a2.multiply(-2);
+    c12 = a1.add(a.add(a3));
+
+    a = a1.multiply(-2);
+    b = a2.multiply(2);
+    c11 = a.add(b);
+
+    c10 = new Point2D(a1.x, a1.y);
+
+    a = b2.multiply(-2);
+    c22 = b1.add(a.add(b3));
+
+    a = b1.multiply(-2);
+    b = b2.multiply(2);
+    c21 = a.add(b);
+
+    c20 = new Point2D(b1.x, b1.y);
+
+    var v0, v1, v2, v3, v4, v5, v6;
+    if ( c12.y === 0 ) {
+        v0 = c12.x*(c10.y - c20.y);
+        v1 = v0 - c11.x*c11.y;
+        v2 = v0 + v1;
+        v3 = c11.y*c11.y;
+
+        poly = new Polynomial(
+            c12.x*c22.y*c22.y,
+            2*c12.x*c21.y*c22.y,
+            c12.x*c21.y*c21.y - c22.x*v3 - c22.y*v0 - c22.y*v1,
+            -c21.x*v3 - c21.y*v0 - c21.y*v1,
+            (c10.x - c20.x)*v3 + (c10.y - c20.y)*v1
+        );
+    } else {
+        v0 = c12.x*c22.y - c12.y*c22.x;
+        v1 = c12.x*c21.y - c21.x*c12.y;
+        v2 = c11.x*c12.y - c11.y*c12.x;
+        v3 = c10.y - c20.y;
+        v4 = c12.y*(c10.x - c20.x) - c12.x*v3;
+        v5 = -c11.y*v2 + c12.y*v4;
+        v6 = v2*v2;
+
+        poly = new Polynomial(
+            v0*v0,
+            2*v0*v1,
+            (-c22.y*v6 + c12.y*v1*v1 + c12.y*v0*v4 + v0*v5) / c12.y,
+            (-c21.y*v6 + c12.y*v1*v4 + v1*v5) / c12.y,
+            (v3*v6 + v4*v5) / c12.y
+        );
+    }
+
+    var roots = poly.getRoots();
+    for ( var i = 0; i < roots.length; i++ ) {
+        var s = roots[i];
+
+        if ( 0 <= s && s <= 1 ) {
+            var xRoots = new Polynomial(
+                c12.x,
+                c11.x,
+                c10.x - c20.x - s*c21.x - s*s*c22.x
+            ).getRoots();
+            var yRoots = new Polynomial(
+                c12.y,
+                c11.y,
+                c10.y - c20.y - s*c21.y - s*s*c22.y
+            ).getRoots();
+
+            if ( xRoots.length > 0 && yRoots.length > 0 ) {
+                var TOLERANCE = 1e-4;
+
+                checkRoots:
+                    for ( var j = 0; j < xRoots.length; j++ ) {
+                        var xRoot = xRoots[j];
+
+                        if ( 0 <= xRoot && xRoot <= 1 ) {
+                            for ( var k = 0; k < yRoots.length; k++ ) {
+                                if ( Math.abs( xRoot - yRoots[k] ) < TOLERANCE ) {
+                                    result.points.push( c22.multiply(s*s).add(c21.multiply(s).add(c20)) );
+                                    break checkRoots;
+                                }
+                            }
+                        }
+                    }
+            }
+        }
+    }
+
+    return result;
+};
+
+
+/**
+ *  intersectBezier2Bezier3
+ *
+ *  @param {Point2D} a1
+ *  @param {Point2D} a2
+ *  @param {Point2D} a3
+ *  @param {Point2D} b1
+ *  @param {Point2D} b2
+ *  @param {Point2D} b3
+ *  @param {Point2D} b4
+ *  @returns {Intersection}
+ */
+module.exports.intersectBezier2Bezier3 = function(a1, a2, a3, b1, b2, b3, b4) {
+    var a, b,c, d;
+    var c12, c11, c10;
+    var c23, c22, c21, c20;
+    var result = new Intersection();
+
+    a = a2.multiply(-2);
+    c12 = a1.add(a.add(a3));
+
+    a = a1.multiply(-2);
+    b = a2.multiply(2);
+    c11 = a.add(b);
+
+    c10 = new Point2D(a1.x, a1.y);
+
+    a = b1.multiply(-1);
+    b = b2.multiply(3);
+    c = b3.multiply(-3);
+    d = a.add(b.add(c.add(b4)));
+    c23 = new Vector2D(d.x, d.y);
+
+    a = b1.multiply(3);
+    b = b2.multiply(-6);
+    c = b3.multiply(3);
+    d = a.add(b.add(c));
+    c22 = new Vector2D(d.x, d.y);
+
+    a = b1.multiply(-3);
+    b = b2.multiply(3);
+    c = a.add(b);
+    c21 = new Vector2D(c.x, c.y);
+
+    c20 = new Vector2D(b1.x, b1.y);
+
+    var c10x2 = c10.x*c10.x;
+    var c10y2 = c10.y*c10.y;
+    var c11x2 = c11.x*c11.x;
+    var c11y2 = c11.y*c11.y;
+    var c12x2 = c12.x*c12.x;
+    var c12y2 = c12.y*c12.y;
+    var c20x2 = c20.x*c20.x;
+    var c20y2 = c20.y*c20.y;
+    var c21x2 = c21.x*c21.x;
+    var c21y2 = c21.y*c21.y;
+    var c22x2 = c22.x*c22.x;
+    var c22y2 = c22.y*c22.y;
+    var c23x2 = c23.x*c23.x;
+    var c23y2 = c23.y*c23.y;
+
+    var poly = new Polynomial(
+        -2*c12.x*c12.y*c23.x*c23.y + c12x2*c23y2 + c12y2*c23x2,
+        -2*c12.x*c12.y*c22.x*c23.y - 2*c12.x*c12.y*c22.y*c23.x + 2*c12y2*c22.x*c23.x +
+            2*c12x2*c22.y*c23.y,
+        -2*c12.x*c21.x*c12.y*c23.y - 2*c12.x*c12.y*c21.y*c23.x - 2*c12.x*c12.y*c22.x*c22.y +
+            2*c21.x*c12y2*c23.x + c12y2*c22x2 + c12x2*(2*c21.y*c23.y + c22y2),
+        2*c10.x*c12.x*c12.y*c23.y + 2*c10.y*c12.x*c12.y*c23.x + c11.x*c11.y*c12.x*c23.y +
+            c11.x*c11.y*c12.y*c23.x - 2*c20.x*c12.x*c12.y*c23.y - 2*c12.x*c20.y*c12.y*c23.x -
+            2*c12.x*c21.x*c12.y*c22.y - 2*c12.x*c12.y*c21.y*c22.x - 2*c10.x*c12y2*c23.x -
+            2*c10.y*c12x2*c23.y + 2*c20.x*c12y2*c23.x + 2*c21.x*c12y2*c22.x -
+            c11y2*c12.x*c23.x - c11x2*c12.y*c23.y + c12x2*(2*c20.y*c23.y + 2*c21.y*c22.y),
+        2*c10.x*c12.x*c12.y*c22.y + 2*c10.y*c12.x*c12.y*c22.x + c11.x*c11.y*c12.x*c22.y +
+            c11.x*c11.y*c12.y*c22.x - 2*c20.x*c12.x*c12.y*c22.y - 2*c12.x*c20.y*c12.y*c22.x -
+            2*c12.x*c21.x*c12.y*c21.y - 2*c10.x*c12y2*c22.x - 2*c10.y*c12x2*c22.y +
+            2*c20.x*c12y2*c22.x - c11y2*c12.x*c22.x - c11x2*c12.y*c22.y + c21x2*c12y2 +
+            c12x2*(2*c20.y*c22.y + c21y2),
+        2*c10.x*c12.x*c12.y*c21.y + 2*c10.y*c12.x*c21.x*c12.y + c11.x*c11.y*c12.x*c21.y +
+            c11.x*c11.y*c21.x*c12.y - 2*c20.x*c12.x*c12.y*c21.y - 2*c12.x*c20.y*c21.x*c12.y -
+            2*c10.x*c21.x*c12y2 - 2*c10.y*c12x2*c21.y + 2*c20.x*c21.x*c12y2 -
+            c11y2*c12.x*c21.x - c11x2*c12.y*c21.y + 2*c12x2*c20.y*c21.y,
+        -2*c10.x*c10.y*c12.x*c12.y - c10.x*c11.x*c11.y*c12.y - c10.y*c11.x*c11.y*c12.x +
+            2*c10.x*c12.x*c20.y*c12.y + 2*c10.y*c20.x*c12.x*c12.y + c11.x*c20.x*c11.y*c12.y +
+            c11.x*c11.y*c12.x*c20.y - 2*c20.x*c12.x*c20.y*c12.y - 2*c10.x*c20.x*c12y2 +
+            c10.x*c11y2*c12.x + c10.y*c11x2*c12.y - 2*c10.y*c12x2*c20.y -
+            c20.x*c11y2*c12.x - c11x2*c20.y*c12.y + c10x2*c12y2 + c10y2*c12x2 +
+            c20x2*c12y2 + c12x2*c20y2
+    );
+    var roots = poly.getRootsInInterval(0,1);
+    module.exports.Utils.removeMultipleRootsIn01(roots);
+
+    for ( var i = 0; i < roots.length; i++ ) {
+        var s = roots[i];
+        var xRoots = new Polynomial(
+            c12.x,
+            c11.x,
+            c10.x - c20.x - s*c21.x - s*s*c22.x - s*s*s*c23.x
+        ).getRoots();
+        var yRoots = new Polynomial(
+            c12.y,
+            c11.y,
+            c10.y - c20.y - s*c21.y - s*s*c22.y - s*s*s*c23.y
+        ).getRoots();
+
+        if ( xRoots.length > 0 && yRoots.length > 0 ) {
+            var TOLERANCE = 1e-4;
+
+            checkRoots:
+                for ( var j = 0; j < xRoots.length; j++ ) {
+                    var xRoot = xRoots[j];
+
+                    if ( 0 <= xRoot && xRoot <= 1 ) {
+                        for ( var k = 0; k < yRoots.length; k++ ) {
+                            if ( Math.abs( xRoot - yRoots[k] ) < TOLERANCE ) {
+                                var v = c23.multiply(s * s * s).add(c22.multiply(s * s).add(c21.multiply(s).add(c20)));
+                                result.points.push(new Point2D(v.x, v.y));
+                                break checkRoots;
+                            }
+                        }
+                    }
+                }
+        }
+    }
+
+    return result;
+
+};
+
+/**
+ *  intersectBezier2Ellipse
+ *
+ *  @param {Point2D} p1
+ *  @param {Point2D} p2
+ *  @param {Point2D} p3
+ *  @param {Point2D} ec
+ *  @param {Number} rx
+ *  @param {Number} ry
+ *  @returns {Intersection}
+ */
+module.exports.intersectBezier2Ellipse = function(p1, p2, p3, ec, rx, ry) {
+    var a, b;       // temporary variables
+    var c2, c1, c0; // coefficients of quadratic
+    var result = new Intersection();
+
+    a = p2.multiply(-2);
+    c2 = p1.add(a.add(p3));
+
+    a = p1.multiply(-2);
+    b = p2.multiply(2);
+    c1 = a.add(b);
+
+    c0 = new Point2D(p1.x, p1.y);
+
+    var rxrx  = rx*rx;
+    var ryry  = ry*ry;
+    var roots = new Polynomial(
+        ryry*c2.x*c2.x + rxrx*c2.y*c2.y,
+        2*(ryry*c2.x*c1.x + rxrx*c2.y*c1.y),
+        ryry*(2*c2.x*c0.x + c1.x*c1.x) + rxrx*(2*c2.y*c0.y+c1.y*c1.y) -
+            2*(ryry*ec.x*c2.x + rxrx*ec.y*c2.y),
+        2*(ryry*c1.x*(c0.x-ec.x) + rxrx*c1.y*(c0.y-ec.y)),
+        ryry*(c0.x*c0.x+ec.x*ec.x) + rxrx*(c0.y*c0.y + ec.y*ec.y) -
+            2*(ryry*ec.x*c0.x + rxrx*ec.y*c0.y) - rxrx*ryry
+    ).getRoots();
+
+    for ( var i = 0; i < roots.length; i++ ) {
+        var t = roots[i];
+
+        if ( 0 <= t && t <= 1 )
+            result.points.push( c2.multiply(t*t).add(c1.multiply(t).add(c0)) );
+    }
+
+    return result;
+};
+
+
+/**
+ *  intersectBezier2Line
+ *
+ *  @param {Point2D} p1
+ *  @param {Point2D} p2
+ *  @param {Point2D} p3
+ *  @param {Point2D} a1
+ *  @param {Point2D} a2
+ *  @returns {Intersection}
+ */
+module.exports.intersectBezier2Line = function(p1, p2, p3, a1, a2) {
+    var a, b;             // temporary variables
+    var c2, c1, c0;       // coefficients of quadratic
+    var cl;               // c coefficient for normal form of line
+    var n;                // normal for normal form of line
+    var min = a1.min(a2); // used to determine if point is on line segment
+    var max = a1.max(a2); // used to determine if point is on line segment
+    var result = new Intersection();
+
+    a = p2.multiply(-2);
+    c2 = p1.add(a.add(p3));
+
+    a = p1.multiply(-2);
+    b = p2.multiply(2);
+    c1 = a.add(b);
+
+    c0 = new Point2D(p1.x, p1.y);
+
+    // Convert line to normal form: ax + by + c = 0
+    // Find normal to line: negative inverse of original line's slope
+    n = new Vector2D(a1.y - a2.y, a2.x - a1.x);
+
+    // Determine new c coefficient
+    cl = a1.x*a2.y - a2.x*a1.y;
+
+    // Transform cubic coefficients to line's coordinate system and find roots
+    // of cubic
+    roots = new Polynomial(
+        n.dot(c2),
+        n.dot(c1),
+        n.dot(c0) + cl
+    ).getRoots();
+
+    // Any roots in closed interval [0,1] are intersections on Bezier, but
+    // might not be on the line segment.
+    // Find intersections and calculate point coordinates
+    for ( var i = 0; i < roots.length; i++ ) {
+        var t = roots[i];
+
+        if ( 0 <= t && t <= 1 ) {
+            // We're within the Bezier curve
+            // Find point on Bezier
+            var p4 = p1.lerp(p2, t);
+            var p5 = p2.lerp(p3, t);
+
+            var p6 = p4.lerp(p5, t);
+
+            // See if point is on line segment
+            // Had to make special cases for vertical and horizontal lines due
+            // to slight errors in calculation of p6
+            if ( a1.x == a2.x ) {
+                if ( min.y <= p6.y && p6.y <= max.y ) {
+                    result.appendPoint( p6 );
+                }
+            } else if ( a1.y == a2.y ) {
+                if ( min.x <= p6.x && p6.x <= max.x ) {
+                    result.appendPoint( p6 );
+                }
+            } else if (min.x <= p6.x && p6.x <= max.x && min.y <= p6.y && p6.y <= max.y) {
+                result.appendPoint( p6 );
+            }
+        }
+    }
+
+    return result;
+};
+
+
+/**
+ *  intersectBezier3Bezier3
+ *
+ *  @param {Point2D} a1
+ *  @param {Point2D} a2
+ *  @param {Point2D} a3
+ *  @param {Point2D} a4
+ *  @param {Point2D} b1
+ *  @param {Point2D} b2
+ *  @param {Point2D} b3
+ *  @param {Point2D} b4
+ *  @returns {Intersection}
+ */
+module.exports.intersectBezier3Bezier3 = function(a1, a2, a3, a4, b1, b2, b3, b4) {
+    var a, b, c, d;         // temporary variables
+    var c13, c12, c11, c10; // coefficients of cubic
+    var c23, c22, c21, c20; // coefficients of cubic
+    var result = new Intersection();
+
+    // Calculate the coefficients of cubic polynomial
+    a = a1.multiply(-1);
+    b = a2.multiply(3);
+    c = a3.multiply(-3);
+    d = a.add(b.add(c.add(a4)));
+    c13 = new Vector2D(d.x, d.y);
+
+    a = a1.multiply(3);
+    b = a2.multiply(-6);
+    c = a3.multiply(3);
+    d = a.add(b.add(c));
+    c12 = new Vector2D(d.x, d.y);
+
+    a = a1.multiply(-3);
+    b = a2.multiply(3);
+    c = a.add(b);
+    c11 = new Vector2D(c.x, c.y);
+
+    c10 = new Vector2D(a1.x, a1.y);
+
+    a = b1.multiply(-1);
+    b = b2.multiply(3);
+    c = b3.multiply(-3);
+    d = a.add(b.add(c.add(b4)));
+    c23 = new Vector2D(d.x, d.y);
+
+    a = b1.multiply(3);
+    b = b2.multiply(-6);
+    c = b3.multiply(3);
+    d = a.add(b.add(c));
+    c22 = new Vector2D(d.x, d.y);
+
+    a = b1.multiply(-3);
+    b = b2.multiply(3);
+    c = a.add(b);
+    c21 = new Vector2D(c.x, c.y);
+
+    c20 = new Vector2D(b1.x, b1.y);
+
+    var c10x2 = c10.x*c10.x;
+    var c10x3 = c10.x*c10.x*c10.x;
+    var c10y2 = c10.y*c10.y;
+    var c10y3 = c10.y*c10.y*c10.y;
+    var c11x2 = c11.x*c11.x;
+    var c11x3 = c11.x*c11.x*c11.x;
+    var c11y2 = c11.y*c11.y;
+    var c11y3 = c11.y*c11.y*c11.y;
+    var c12x2 = c12.x*c12.x;
+    var c12x3 = c12.x*c12.x*c12.x;
+    var c12y2 = c12.y*c12.y;
+    var c12y3 = c12.y*c12.y*c12.y;
+    var c13x2 = c13.x*c13.x;
+    var c13x3 = c13.x*c13.x*c13.x;
+    var c13y2 = c13.y*c13.y;
+    var c13y3 = c13.y*c13.y*c13.y;
+    var c20x2 = c20.x*c20.x;
+    var c20x3 = c20.x*c20.x*c20.x;
+    var c20y2 = c20.y*c20.y;
+    var c20y3 = c20.y*c20.y*c20.y;
+    var c21x2 = c21.x*c21.x;
+    var c21x3 = c21.x*c21.x*c21.x;
+    var c21y2 = c21.y*c21.y;
+    var c22x2 = c22.x*c22.x;
+    var c22x3 = c22.x*c22.x*c22.x;
+    var c22y2 = c22.y*c22.y;
+    var c23x2 = c23.x*c23.x;
+    var c23x3 = c23.x*c23.x*c23.x;
+    var c23y2 = c23.y*c23.y;
+    var c23y3 = c23.y*c23.y*c23.y;
+    var poly = new Polynomial(
+        -c13x3*c23y3 + c13y3*c23x3 - 3*c13.x*c13y2*c23x2*c23.y +
+            3*c13x2*c13.y*c23.x*c23y2,
+        -6*c13.x*c22.x*c13y2*c23.x*c23.y + 6*c13x2*c13.y*c22.y*c23.x*c23.y + 3*c22.x*c13y3*c23x2 -
+            3*c13x3*c22.y*c23y2 - 3*c13.x*c13y2*c22.y*c23x2 + 3*c13x2*c22.x*c13.y*c23y2,
+        -6*c21.x*c13.x*c13y2*c23.x*c23.y - 6*c13.x*c22.x*c13y2*c22.y*c23.x + 6*c13x2*c22.x*c13.y*c22.y*c23.y +
+            3*c21.x*c13y3*c23x2 + 3*c22x2*c13y3*c23.x + 3*c21.x*c13x2*c13.y*c23y2 - 3*c13.x*c21.y*c13y2*c23x2 -
+            3*c13.x*c22x2*c13y2*c23.y + c13x2*c13.y*c23.x*(6*c21.y*c23.y + 3*c22y2) + c13x3*(-c21.y*c23y2 -
+            2*c22y2*c23.y - c23.y*(2*c21.y*c23.y + c22y2)),
+        c11.x*c12.y*c13.x*c13.y*c23.x*c23.y - c11.y*c12.x*c13.x*c13.y*c23.x*c23.y + 6*c21.x*c22.x*c13y3*c23.x +
+            3*c11.x*c12.x*c13.x*c13.y*c23y2 + 6*c10.x*c13.x*c13y2*c23.x*c23.y - 3*c11.x*c12.x*c13y2*c23.x*c23.y -
+            3*c11.y*c12.y*c13.x*c13.y*c23x2 - 6*c10.y*c13x2*c13.y*c23.x*c23.y - 6*c20.x*c13.x*c13y2*c23.x*c23.y +
+            3*c11.y*c12.y*c13x2*c23.x*c23.y - 2*c12.x*c12y2*c13.x*c23.x*c23.y - 6*c21.x*c13.x*c22.x*c13y2*c23.y -
+            6*c21.x*c13.x*c13y2*c22.y*c23.x - 6*c13.x*c21.y*c22.x*c13y2*c23.x + 6*c21.x*c13x2*c13.y*c22.y*c23.y +
+            2*c12x2*c12.y*c13.y*c23.x*c23.y + c22x3*c13y3 - 3*c10.x*c13y3*c23x2 + 3*c10.y*c13x3*c23y2 +
+            3*c20.x*c13y3*c23x2 + c12y3*c13.x*c23x2 - c12x3*c13.y*c23y2 - 3*c10.x*c13x2*c13.y*c23y2 +
+            3*c10.y*c13.x*c13y2*c23x2 - 2*c11.x*c12.y*c13x2*c23y2 + c11.x*c12.y*c13y2*c23x2 - c11.y*c12.x*c13x2*c23y2 +
+            2*c11.y*c12.x*c13y2*c23x2 + 3*c20.x*c13x2*c13.y*c23y2 - c12.x*c12y2*c13.y*c23x2 -
+            3*c20.y*c13.x*c13y2*c23x2 + c12x2*c12.y*c13.x*c23y2 - 3*c13.x*c22x2*c13y2*c22.y +
+            c13x2*c13.y*c23.x*(6*c20.y*c23.y + 6*c21.y*c22.y) + c13x2*c22.x*c13.y*(6*c21.y*c23.y + 3*c22y2) +
+            c13x3*(-2*c21.y*c22.y*c23.y - c20.y*c23y2 - c22.y*(2*c21.y*c23.y + c22y2) - c23.y*(2*c20.y*c23.y + 2*c21.y*c22.y)),
+        6*c11.x*c12.x*c13.x*c13.y*c22.y*c23.y + c11.x*c12.y*c13.x*c22.x*c13.y*c23.y + c11.x*c12.y*c13.x*c13.y*c22.y*c23.x -
+            c11.y*c12.x*c13.x*c22.x*c13.y*c23.y - c11.y*c12.x*c13.x*c13.y*c22.y*c23.x - 6*c11.y*c12.y*c13.x*c22.x*c13.y*c23.x -
+            6*c10.x*c22.x*c13y3*c23.x + 6*c20.x*c22.x*c13y3*c23.x + 6*c10.y*c13x3*c22.y*c23.y + 2*c12y3*c13.x*c22.x*c23.x -
+            2*c12x3*c13.y*c22.y*c23.y + 6*c10.x*c13.x*c22.x*c13y2*c23.y + 6*c10.x*c13.x*c13y2*c22.y*c23.x +
+            6*c10.y*c13.x*c22.x*c13y2*c23.x - 3*c11.x*c12.x*c22.x*c13y2*c23.y - 3*c11.x*c12.x*c13y2*c22.y*c23.x +
+            2*c11.x*c12.y*c22.x*c13y2*c23.x + 4*c11.y*c12.x*c22.x*c13y2*c23.x - 6*c10.x*c13x2*c13.y*c22.y*c23.y -
+            6*c10.y*c13x2*c22.x*c13.y*c23.y - 6*c10.y*c13x2*c13.y*c22.y*c23.x - 4*c11.x*c12.y*c13x2*c22.y*c23.y -
+            6*c20.x*c13.x*c22.x*c13y2*c23.y - 6*c20.x*c13.x*c13y2*c22.y*c23.x - 2*c11.y*c12.x*c13x2*c22.y*c23.y +
+            3*c11.y*c12.y*c13x2*c22.x*c23.y + 3*c11.y*c12.y*c13x2*c22.y*c23.x - 2*c12.x*c12y2*c13.x*c22.x*c23.y -
+            2*c12.x*c12y2*c13.x*c22.y*c23.x - 2*c12.x*c12y2*c22.x*c13.y*c23.x - 6*c20.y*c13.x*c22.x*c13y2*c23.x -
+            6*c21.x*c13.x*c21.y*c13y2*c23.x - 6*c21.x*c13.x*c22.x*c13y2*c22.y + 6*c20.x*c13x2*c13.y*c22.y*c23.y +
+            2*c12x2*c12.y*c13.x*c22.y*c23.y + 2*c12x2*c12.y*c22.x*c13.y*c23.y + 2*c12x2*c12.y*c13.y*c22.y*c23.x +
+            3*c21.x*c22x2*c13y3 + 3*c21x2*c13y3*c23.x - 3*c13.x*c21.y*c22x2*c13y2 - 3*c21x2*c13.x*c13y2*c23.y +
+            c13x2*c22.x*c13.y*(6*c20.y*c23.y + 6*c21.y*c22.y) + c13x2*c13.y*c23.x*(6*c20.y*c22.y + 3*c21y2) +
+            c21.x*c13x2*c13.y*(6*c21.y*c23.y + 3*c22y2) + c13x3*(-2*c20.y*c22.y*c23.y - c23.y*(2*c20.y*c22.y + c21y2) -
+            c21.y*(2*c21.y*c23.y + c22y2) - c22.y*(2*c20.y*c23.y + 2*c21.y*c22.y)),
+        c11.x*c21.x*c12.y*c13.x*c13.y*c23.y + c11.x*c12.y*c13.x*c21.y*c13.y*c23.x + c11.x*c12.y*c13.x*c22.x*c13.y*c22.y -
+            c11.y*c12.x*c21.x*c13.x*c13.y*c23.y - c11.y*c12.x*c13.x*c21.y*c13.y*c23.x - c11.y*c12.x*c13.x*c22.x*c13.y*c22.y -
+            6*c11.y*c21.x*c12.y*c13.x*c13.y*c23.x - 6*c10.x*c21.x*c13y3*c23.x + 6*c20.x*c21.x*c13y3*c23.x +
+            2*c21.x*c12y3*c13.x*c23.x + 6*c10.x*c21.x*c13.x*c13y2*c23.y + 6*c10.x*c13.x*c21.y*c13y2*c23.x +
+            6*c10.x*c13.x*c22.x*c13y2*c22.y + 6*c10.y*c21.x*c13.x*c13y2*c23.x - 3*c11.x*c12.x*c21.x*c13y2*c23.y -
+            3*c11.x*c12.x*c21.y*c13y2*c23.x - 3*c11.x*c12.x*c22.x*c13y2*c22.y + 2*c11.x*c21.x*c12.y*c13y2*c23.x +
+            4*c11.y*c12.x*c21.x*c13y2*c23.x - 6*c10.y*c21.x*c13x2*c13.y*c23.y - 6*c10.y*c13x2*c21.y*c13.y*c23.x -
+            6*c10.y*c13x2*c22.x*c13.y*c22.y - 6*c20.x*c21.x*c13.x*c13y2*c23.y - 6*c20.x*c13.x*c21.y*c13y2*c23.x -
+            6*c20.x*c13.x*c22.x*c13y2*c22.y + 3*c11.y*c21.x*c12.y*c13x2*c23.y - 3*c11.y*c12.y*c13.x*c22x2*c13.y +
+            3*c11.y*c12.y*c13x2*c21.y*c23.x + 3*c11.y*c12.y*c13x2*c22.x*c22.y - 2*c12.x*c21.x*c12y2*c13.x*c23.y -
+            2*c12.x*c21.x*c12y2*c13.y*c23.x - 2*c12.x*c12y2*c13.x*c21.y*c23.x - 2*c12.x*c12y2*c13.x*c22.x*c22.y -
+            6*c20.y*c21.x*c13.x*c13y2*c23.x - 6*c21.x*c13.x*c21.y*c22.x*c13y2 + 6*c20.y*c13x2*c21.y*c13.y*c23.x +
+            2*c12x2*c21.x*c12.y*c13.y*c23.y + 2*c12x2*c12.y*c21.y*c13.y*c23.x + 2*c12x2*c12.y*c22.x*c13.y*c22.y -
+            3*c10.x*c22x2*c13y3 + 3*c20.x*c22x2*c13y3 + 3*c21x2*c22.x*c13y3 + c12y3*c13.x*c22x2 +
+            3*c10.y*c13.x*c22x2*c13y2 + c11.x*c12.y*c22x2*c13y2 + 2*c11.y*c12.x*c22x2*c13y2 -
+            c12.x*c12y2*c22x2*c13.y - 3*c20.y*c13.x*c22x2*c13y2 - 3*c21x2*c13.x*c13y2*c22.y +
+            c12x2*c12.y*c13.x*(2*c21.y*c23.y + c22y2) + c11.x*c12.x*c13.x*c13.y*(6*c21.y*c23.y + 3*c22y2) +
+            c21.x*c13x2*c13.y*(6*c20.y*c23.y + 6*c21.y*c22.y) + c12x3*c13.y*(-2*c21.y*c23.y - c22y2) +
+            c10.y*c13x3*(6*c21.y*c23.y + 3*c22y2) + c11.y*c12.x*c13x2*(-2*c21.y*c23.y - c22y2) +
+            c11.x*c12.y*c13x2*(-4*c21.y*c23.y - 2*c22y2) + c10.x*c13x2*c13.y*(-6*c21.y*c23.y - 3*c22y2) +
+            c13x2*c22.x*c13.y*(6*c20.y*c22.y + 3*c21y2) + c20.x*c13x2*c13.y*(6*c21.y*c23.y + 3*c22y2) +
+            c13x3*(-2*c20.y*c21.y*c23.y - c22.y*(2*c20.y*c22.y + c21y2) - c20.y*(2*c21.y*c23.y + c22y2) -
+            c21.y*(2*c20.y*c23.y + 2*c21.y*c22.y)),
+        -c10.x*c11.x*c12.y*c13.x*c13.y*c23.y + c10.x*c11.y*c12.x*c13.x*c13.y*c23.y + 6*c10.x*c11.y*c12.y*c13.x*c13.y*c23.x -
+            6*c10.y*c11.x*c12.x*c13.x*c13.y*c23.y - c10.y*c11.x*c12.y*c13.x*c13.y*c23.x + c10.y*c11.y*c12.x*c13.x*c13.y*c23.x +
+            c11.x*c11.y*c12.x*c12.y*c13.x*c23.y - c11.x*c11.y*c12.x*c12.y*c13.y*c23.x + c11.x*c20.x*c12.y*c13.x*c13.y*c23.y +
+            c11.x*c20.y*c12.y*c13.x*c13.y*c23.x + c11.x*c21.x*c12.y*c13.x*c13.y*c22.y + c11.x*c12.y*c13.x*c21.y*c22.x*c13.y -
+            c20.x*c11.y*c12.x*c13.x*c13.y*c23.y - 6*c20.x*c11.y*c12.y*c13.x*c13.y*c23.x - c11.y*c12.x*c20.y*c13.x*c13.y*c23.x -
+            c11.y*c12.x*c21.x*c13.x*c13.y*c22.y - c11.y*c12.x*c13.x*c21.y*c22.x*c13.y - 6*c11.y*c21.x*c12.y*c13.x*c22.x*c13.y -
+            6*c10.x*c20.x*c13y3*c23.x - 6*c10.x*c21.x*c22.x*c13y3 - 2*c10.x*c12y3*c13.x*c23.x + 6*c20.x*c21.x*c22.x*c13y3 +
+            2*c20.x*c12y3*c13.x*c23.x + 2*c21.x*c12y3*c13.x*c22.x + 2*c10.y*c12x3*c13.y*c23.y - 6*c10.x*c10.y*c13.x*c13y2*c23.x +
+            3*c10.x*c11.x*c12.x*c13y2*c23.y - 2*c10.x*c11.x*c12.y*c13y2*c23.x - 4*c10.x*c11.y*c12.x*c13y2*c23.x +
+            3*c10.y*c11.x*c12.x*c13y2*c23.x + 6*c10.x*c10.y*c13x2*c13.y*c23.y + 6*c10.x*c20.x*c13.x*c13y2*c23.y -
+            3*c10.x*c11.y*c12.y*c13x2*c23.y + 2*c10.x*c12.x*c12y2*c13.x*c23.y + 2*c10.x*c12.x*c12y2*c13.y*c23.x +
+            6*c10.x*c20.y*c13.x*c13y2*c23.x + 6*c10.x*c21.x*c13.x*c13y2*c22.y + 6*c10.x*c13.x*c21.y*c22.x*c13y2 +
+            4*c10.y*c11.x*c12.y*c13x2*c23.y + 6*c10.y*c20.x*c13.x*c13y2*c23.x + 2*c10.y*c11.y*c12.x*c13x2*c23.y -
+            3*c10.y*c11.y*c12.y*c13x2*c23.x + 2*c10.y*c12.x*c12y2*c13.x*c23.x + 6*c10.y*c21.x*c13.x*c22.x*c13y2 -
+            3*c11.x*c20.x*c12.x*c13y2*c23.y + 2*c11.x*c20.x*c12.y*c13y2*c23.x + c11.x*c11.y*c12y2*c13.x*c23.x -
+            3*c11.x*c12.x*c20.y*c13y2*c23.x - 3*c11.x*c12.x*c21.x*c13y2*c22.y - 3*c11.x*c12.x*c21.y*c22.x*c13y2 +
+            2*c11.x*c21.x*c12.y*c22.x*c13y2 + 4*c20.x*c11.y*c12.x*c13y2*c23.x + 4*c11.y*c12.x*c21.x*c22.x*c13y2 -
+            2*c10.x*c12x2*c12.y*c13.y*c23.y - 6*c10.y*c20.x*c13x2*c13.y*c23.y - 6*c10.y*c20.y*c13x2*c13.y*c23.x -
+            6*c10.y*c21.x*c13x2*c13.y*c22.y - 2*c10.y*c12x2*c12.y*c13.x*c23.y - 2*c10.y*c12x2*c12.y*c13.y*c23.x -
+            6*c10.y*c13x2*c21.y*c22.x*c13.y - c11.x*c11.y*c12x2*c13.y*c23.y - 2*c11.x*c11y2*c13.x*c13.y*c23.x +
+            3*c20.x*c11.y*c12.y*c13x2*c23.y - 2*c20.x*c12.x*c12y2*c13.x*c23.y - 2*c20.x*c12.x*c12y2*c13.y*c23.x -
+            6*c20.x*c20.y*c13.x*c13y2*c23.x - 6*c20.x*c21.x*c13.x*c13y2*c22.y - 6*c20.x*c13.x*c21.y*c22.x*c13y2 +
+            3*c11.y*c20.y*c12.y*c13x2*c23.x + 3*c11.y*c21.x*c12.y*c13x2*c22.y + 3*c11.y*c12.y*c13x2*c21.y*c22.x -
+            2*c12.x*c20.y*c12y2*c13.x*c23.x - 2*c12.x*c21.x*c12y2*c13.x*c22.y - 2*c12.x*c21.x*c12y2*c22.x*c13.y -
+            2*c12.x*c12y2*c13.x*c21.y*c22.x - 6*c20.y*c21.x*c13.x*c22.x*c13y2 - c11y2*c12.x*c12.y*c13.x*c23.x +
+            2*c20.x*c12x2*c12.y*c13.y*c23.y + 6*c20.y*c13x2*c21.y*c22.x*c13.y + 2*c11x2*c11.y*c13.x*c13.y*c23.y +
+            c11x2*c12.x*c12.y*c13.y*c23.y + 2*c12x2*c20.y*c12.y*c13.y*c23.x + 2*c12x2*c21.x*c12.y*c13.y*c22.y +
+            2*c12x2*c12.y*c21.y*c22.x*c13.y + c21x3*c13y3 + 3*c10x2*c13y3*c23.x - 3*c10y2*c13x3*c23.y +
+            3*c20x2*c13y3*c23.x + c11y3*c13x2*c23.x - c11x3*c13y2*c23.y - c11.x*c11y2*c13x2*c23.y +
+            c11x2*c11.y*c13y2*c23.x - 3*c10x2*c13.x*c13y2*c23.y + 3*c10y2*c13x2*c13.y*c23.x - c11x2*c12y2*c13.x*c23.y +
+            c11y2*c12x2*c13.y*c23.x - 3*c21x2*c13.x*c21.y*c13y2 - 3*c20x2*c13.x*c13y2*c23.y + 3*c20y2*c13x2*c13.y*c23.x +
+            c11.x*c12.x*c13.x*c13.y*(6*c20.y*c23.y + 6*c21.y*c22.y) + c12x3*c13.y*(-2*c20.y*c23.y - 2*c21.y*c22.y) +
+            c10.y*c13x3*(6*c20.y*c23.y + 6*c21.y*c22.y) + c11.y*c12.x*c13x2*(-2*c20.y*c23.y - 2*c21.y*c22.y) +
+            c12x2*c12.y*c13.x*(2*c20.y*c23.y + 2*c21.y*c22.y) + c11.x*c12.y*c13x2*(-4*c20.y*c23.y - 4*c21.y*c22.y) +
+            c10.x*c13x2*c13.y*(-6*c20.y*c23.y - 6*c21.y*c22.y) + c20.x*c13x2*c13.y*(6*c20.y*c23.y + 6*c21.y*c22.y) +
+            c21.x*c13x2*c13.y*(6*c20.y*c22.y + 3*c21y2) + c13x3*(-2*c20.y*c21.y*c22.y - c20y2*c23.y -
+            c21.y*(2*c20.y*c22.y + c21y2) - c20.y*(2*c20.y*c23.y + 2*c21.y*c22.y)),
+        -c10.x*c11.x*c12.y*c13.x*c13.y*c22.y + c10.x*c11.y*c12.x*c13.x*c13.y*c22.y + 6*c10.x*c11.y*c12.y*c13.x*c22.x*c13.y -
+            6*c10.y*c11.x*c12.x*c13.x*c13.y*c22.y - c10.y*c11.x*c12.y*c13.x*c22.x*c13.y + c10.y*c11.y*c12.x*c13.x*c22.x*c13.y +
+            c11.x*c11.y*c12.x*c12.y*c13.x*c22.y - c11.x*c11.y*c12.x*c12.y*c22.x*c13.y + c11.x*c20.x*c12.y*c13.x*c13.y*c22.y +
+            c11.x*c20.y*c12.y*c13.x*c22.x*c13.y + c11.x*c21.x*c12.y*c13.x*c21.y*c13.y - c20.x*c11.y*c12.x*c13.x*c13.y*c22.y -
+            6*c20.x*c11.y*c12.y*c13.x*c22.x*c13.y - c11.y*c12.x*c20.y*c13.x*c22.x*c13.y - c11.y*c12.x*c21.x*c13.x*c21.y*c13.y -
+            6*c10.x*c20.x*c22.x*c13y3 - 2*c10.x*c12y3*c13.x*c22.x + 2*c20.x*c12y3*c13.x*c22.x + 2*c10.y*c12x3*c13.y*c22.y -
+            6*c10.x*c10.y*c13.x*c22.x*c13y2 + 3*c10.x*c11.x*c12.x*c13y2*c22.y - 2*c10.x*c11.x*c12.y*c22.x*c13y2 -
+            4*c10.x*c11.y*c12.x*c22.x*c13y2 + 3*c10.y*c11.x*c12.x*c22.x*c13y2 + 6*c10.x*c10.y*c13x2*c13.y*c22.y +
+            6*c10.x*c20.x*c13.x*c13y2*c22.y - 3*c10.x*c11.y*c12.y*c13x2*c22.y + 2*c10.x*c12.x*c12y2*c13.x*c22.y +
+            2*c10.x*c12.x*c12y2*c22.x*c13.y + 6*c10.x*c20.y*c13.x*c22.x*c13y2 + 6*c10.x*c21.x*c13.x*c21.y*c13y2 +
+            4*c10.y*c11.x*c12.y*c13x2*c22.y + 6*c10.y*c20.x*c13.x*c22.x*c13y2 + 2*c10.y*c11.y*c12.x*c13x2*c22.y -
+            3*c10.y*c11.y*c12.y*c13x2*c22.x + 2*c10.y*c12.x*c12y2*c13.x*c22.x - 3*c11.x*c20.x*c12.x*c13y2*c22.y +
+            2*c11.x*c20.x*c12.y*c22.x*c13y2 + c11.x*c11.y*c12y2*c13.x*c22.x - 3*c11.x*c12.x*c20.y*c22.x*c13y2 -
+            3*c11.x*c12.x*c21.x*c21.y*c13y2 + 4*c20.x*c11.y*c12.x*c22.x*c13y2 - 2*c10.x*c12x2*c12.y*c13.y*c22.y -
+            6*c10.y*c20.x*c13x2*c13.y*c22.y - 6*c10.y*c20.y*c13x2*c22.x*c13.y - 6*c10.y*c21.x*c13x2*c21.y*c13.y -
+            2*c10.y*c12x2*c12.y*c13.x*c22.y - 2*c10.y*c12x2*c12.y*c22.x*c13.y - c11.x*c11.y*c12x2*c13.y*c22.y -
+            2*c11.x*c11y2*c13.x*c22.x*c13.y + 3*c20.x*c11.y*c12.y*c13x2*c22.y - 2*c20.x*c12.x*c12y2*c13.x*c22.y -
+            2*c20.x*c12.x*c12y2*c22.x*c13.y - 6*c20.x*c20.y*c13.x*c22.x*c13y2 - 6*c20.x*c21.x*c13.x*c21.y*c13y2 +
+            3*c11.y*c20.y*c12.y*c13x2*c22.x + 3*c11.y*c21.x*c12.y*c13x2*c21.y - 2*c12.x*c20.y*c12y2*c13.x*c22.x -
+            2*c12.x*c21.x*c12y2*c13.x*c21.y - c11y2*c12.x*c12.y*c13.x*c22.x + 2*c20.x*c12x2*c12.y*c13.y*c22.y -
+            3*c11.y*c21x2*c12.y*c13.x*c13.y + 6*c20.y*c21.x*c13x2*c21.y*c13.y + 2*c11x2*c11.y*c13.x*c13.y*c22.y +
+            c11x2*c12.x*c12.y*c13.y*c22.y + 2*c12x2*c20.y*c12.y*c22.x*c13.y + 2*c12x2*c21.x*c12.y*c21.y*c13.y -
+            3*c10.x*c21x2*c13y3 + 3*c20.x*c21x2*c13y3 + 3*c10x2*c22.x*c13y3 - 3*c10y2*c13x3*c22.y + 3*c20x2*c22.x*c13y3 +
+            c21x2*c12y3*c13.x + c11y3*c13x2*c22.x - c11x3*c13y2*c22.y + 3*c10.y*c21x2*c13.x*c13y2 -
+            c11.x*c11y2*c13x2*c22.y + c11.x*c21x2*c12.y*c13y2 + 2*c11.y*c12.x*c21x2*c13y2 + c11x2*c11.y*c22.x*c13y2 -
+            c12.x*c21x2*c12y2*c13.y - 3*c20.y*c21x2*c13.x*c13y2 - 3*c10x2*c13.x*c13y2*c22.y + 3*c10y2*c13x2*c22.x*c13.y -
+            c11x2*c12y2*c13.x*c22.y + c11y2*c12x2*c22.x*c13.y - 3*c20x2*c13.x*c13y2*c22.y + 3*c20y2*c13x2*c22.x*c13.y +
+            c12x2*c12.y*c13.x*(2*c20.y*c22.y + c21y2) + c11.x*c12.x*c13.x*c13.y*(6*c20.y*c22.y + 3*c21y2) +
+            c12x3*c13.y*(-2*c20.y*c22.y - c21y2) + c10.y*c13x3*(6*c20.y*c22.y + 3*c21y2) +
+            c11.y*c12.x*c13x2*(-2*c20.y*c22.y - c21y2) + c11.x*c12.y*c13x2*(-4*c20.y*c22.y - 2*c21y2) +
+            c10.x*c13x2*c13.y*(-6*c20.y*c22.y - 3*c21y2) + c20.x*c13x2*c13.y*(6*c20.y*c22.y + 3*c21y2) +
+            c13x3*(-2*c20.y*c21y2 - c20y2*c22.y - c20.y*(2*c20.y*c22.y + c21y2)),
+        -c10.x*c11.x*c12.y*c13.x*c21.y*c13.y + c10.x*c11.y*c12.x*c13.x*c21.y*c13.y + 6*c10.x*c11.y*c21.x*c12.y*c13.x*c13.y -
+            6*c10.y*c11.x*c12.x*c13.x*c21.y*c13.y - c10.y*c11.x*c21.x*c12.y*c13.x*c13.y + c10.y*c11.y*c12.x*c21.x*c13.x*c13.y -
+            c11.x*c11.y*c12.x*c21.x*c12.y*c13.y + c11.x*c11.y*c12.x*c12.y*c13.x*c21.y + c11.x*c20.x*c12.y*c13.x*c21.y*c13.y +
+            6*c11.x*c12.x*c20.y*c13.x*c21.y*c13.y + c11.x*c20.y*c21.x*c12.y*c13.x*c13.y - c20.x*c11.y*c12.x*c13.x*c21.y*c13.y -
+            6*c20.x*c11.y*c21.x*c12.y*c13.x*c13.y - c11.y*c12.x*c20.y*c21.x*c13.x*c13.y - 6*c10.x*c20.x*c21.x*c13y3 -
+            2*c10.x*c21.x*c12y3*c13.x + 6*c10.y*c20.y*c13x3*c21.y + 2*c20.x*c21.x*c12y3*c13.x + 2*c10.y*c12x3*c21.y*c13.y -
+            2*c12x3*c20.y*c21.y*c13.y - 6*c10.x*c10.y*c21.x*c13.x*c13y2 + 3*c10.x*c11.x*c12.x*c21.y*c13y2 -
+            2*c10.x*c11.x*c21.x*c12.y*c13y2 - 4*c10.x*c11.y*c12.x*c21.x*c13y2 + 3*c10.y*c11.x*c12.x*c21.x*c13y2 +
+            6*c10.x*c10.y*c13x2*c21.y*c13.y + 6*c10.x*c20.x*c13.x*c21.y*c13y2 - 3*c10.x*c11.y*c12.y*c13x2*c21.y +
+            2*c10.x*c12.x*c21.x*c12y2*c13.y + 2*c10.x*c12.x*c12y2*c13.x*c21.y + 6*c10.x*c20.y*c21.x*c13.x*c13y2 +
+            4*c10.y*c11.x*c12.y*c13x2*c21.y + 6*c10.y*c20.x*c21.x*c13.x*c13y2 + 2*c10.y*c11.y*c12.x*c13x2*c21.y -
+            3*c10.y*c11.y*c21.x*c12.y*c13x2 + 2*c10.y*c12.x*c21.x*c12y2*c13.x - 3*c11.x*c20.x*c12.x*c21.y*c13y2 +
+            2*c11.x*c20.x*c21.x*c12.y*c13y2 + c11.x*c11.y*c21.x*c12y2*c13.x - 3*c11.x*c12.x*c20.y*c21.x*c13y2 +
+            4*c20.x*c11.y*c12.x*c21.x*c13y2 - 6*c10.x*c20.y*c13x2*c21.y*c13.y - 2*c10.x*c12x2*c12.y*c21.y*c13.y -
+            6*c10.y*c20.x*c13x2*c21.y*c13.y - 6*c10.y*c20.y*c21.x*c13x2*c13.y - 2*c10.y*c12x2*c21.x*c12.y*c13.y -
+            2*c10.y*c12x2*c12.y*c13.x*c21.y - c11.x*c11.y*c12x2*c21.y*c13.y - 4*c11.x*c20.y*c12.y*c13x2*c21.y -
+            2*c11.x*c11y2*c21.x*c13.x*c13.y + 3*c20.x*c11.y*c12.y*c13x2*c21.y - 2*c20.x*c12.x*c21.x*c12y2*c13.y -
+            2*c20.x*c12.x*c12y2*c13.x*c21.y - 6*c20.x*c20.y*c21.x*c13.x*c13y2 - 2*c11.y*c12.x*c20.y*c13x2*c21.y +
+            3*c11.y*c20.y*c21.x*c12.y*c13x2 - 2*c12.x*c20.y*c21.x*c12y2*c13.x - c11y2*c12.x*c21.x*c12.y*c13.x +
+            6*c20.x*c20.y*c13x2*c21.y*c13.y + 2*c20.x*c12x2*c12.y*c21.y*c13.y + 2*c11x2*c11.y*c13.x*c21.y*c13.y +
+            c11x2*c12.x*c12.y*c21.y*c13.y + 2*c12x2*c20.y*c21.x*c12.y*c13.y + 2*c12x2*c20.y*c12.y*c13.x*c21.y +
+            3*c10x2*c21.x*c13y3 - 3*c10y2*c13x3*c21.y + 3*c20x2*c21.x*c13y3 + c11y3*c21.x*c13x2 - c11x3*c21.y*c13y2 -
+            3*c20y2*c13x3*c21.y - c11.x*c11y2*c13x2*c21.y + c11x2*c11.y*c21.x*c13y2 - 3*c10x2*c13.x*c21.y*c13y2 +
+            3*c10y2*c21.x*c13x2*c13.y - c11x2*c12y2*c13.x*c21.y + c11y2*c12x2*c21.x*c13.y - 3*c20x2*c13.x*c21.y*c13y2 +
+            3*c20y2*c21.x*c13x2*c13.y,
+        c10.x*c10.y*c11.x*c12.y*c13.x*c13.y - c10.x*c10.y*c11.y*c12.x*c13.x*c13.y + c10.x*c11.x*c11.y*c12.x*c12.y*c13.y -
+            c10.y*c11.x*c11.y*c12.x*c12.y*c13.x - c10.x*c11.x*c20.y*c12.y*c13.x*c13.y + 6*c10.x*c20.x*c11.y*c12.y*c13.x*c13.y +
+            c10.x*c11.y*c12.x*c20.y*c13.x*c13.y - c10.y*c11.x*c20.x*c12.y*c13.x*c13.y - 6*c10.y*c11.x*c12.x*c20.y*c13.x*c13.y +
+            c10.y*c20.x*c11.y*c12.x*c13.x*c13.y - c11.x*c20.x*c11.y*c12.x*c12.y*c13.y + c11.x*c11.y*c12.x*c20.y*c12.y*c13.x +
+            c11.x*c20.x*c20.y*c12.y*c13.x*c13.y - c20.x*c11.y*c12.x*c20.y*c13.x*c13.y - 2*c10.x*c20.x*c12y3*c13.x +
+            2*c10.y*c12x3*c20.y*c13.y - 3*c10.x*c10.y*c11.x*c12.x*c13y2 - 6*c10.x*c10.y*c20.x*c13.x*c13y2 +
+            3*c10.x*c10.y*c11.y*c12.y*c13x2 - 2*c10.x*c10.y*c12.x*c12y2*c13.x - 2*c10.x*c11.x*c20.x*c12.y*c13y2 -
+            c10.x*c11.x*c11.y*c12y2*c13.x + 3*c10.x*c11.x*c12.x*c20.y*c13y2 - 4*c10.x*c20.x*c11.y*c12.x*c13y2 +
+            3*c10.y*c11.x*c20.x*c12.x*c13y2 + 6*c10.x*c10.y*c20.y*c13x2*c13.y + 2*c10.x*c10.y*c12x2*c12.y*c13.y +
+            2*c10.x*c11.x*c11y2*c13.x*c13.y + 2*c10.x*c20.x*c12.x*c12y2*c13.y + 6*c10.x*c20.x*c20.y*c13.x*c13y2 -
+            3*c10.x*c11.y*c20.y*c12.y*c13x2 + 2*c10.x*c12.x*c20.y*c12y2*c13.x + c10.x*c11y2*c12.x*c12.y*c13.x +
+            c10.y*c11.x*c11.y*c12x2*c13.y + 4*c10.y*c11.x*c20.y*c12.y*c13x2 - 3*c10.y*c20.x*c11.y*c12.y*c13x2 +
+            2*c10.y*c20.x*c12.x*c12y2*c13.x + 2*c10.y*c11.y*c12.x*c20.y*c13x2 + c11.x*c20.x*c11.y*c12y2*c13.x -
+            3*c11.x*c20.x*c12.x*c20.y*c13y2 - 2*c10.x*c12x2*c20.y*c12.y*c13.y - 6*c10.y*c20.x*c20.y*c13x2*c13.y -
+            2*c10.y*c20.x*c12x2*c12.y*c13.y - 2*c10.y*c11x2*c11.y*c13.x*c13.y - c10.y*c11x2*c12.x*c12.y*c13.y -
+            2*c10.y*c12x2*c20.y*c12.y*c13.x - 2*c11.x*c20.x*c11y2*c13.x*c13.y - c11.x*c11.y*c12x2*c20.y*c13.y +
+            3*c20.x*c11.y*c20.y*c12.y*c13x2 - 2*c20.x*c12.x*c20.y*c12y2*c13.x - c20.x*c11y2*c12.x*c12.y*c13.x +
+            3*c10y2*c11.x*c12.x*c13.x*c13.y + 3*c11.x*c12.x*c20y2*c13.x*c13.y + 2*c20.x*c12x2*c20.y*c12.y*c13.y -
+            3*c10x2*c11.y*c12.y*c13.x*c13.y + 2*c11x2*c11.y*c20.y*c13.x*c13.y + c11x2*c12.x*c20.y*c12.y*c13.y -
+            3*c20x2*c11.y*c12.y*c13.x*c13.y - c10x3*c13y3 + c10y3*c13x3 + c20x3*c13y3 - c20y3*c13x3 -
+            3*c10.x*c20x2*c13y3 - c10.x*c11y3*c13x2 + 3*c10x2*c20.x*c13y3 + c10.y*c11x3*c13y2 +
+            3*c10.y*c20y2*c13x3 + c20.x*c11y3*c13x2 + c10x2*c12y3*c13.x - 3*c10y2*c20.y*c13x3 - c10y2*c12x3*c13.y +
+            c20x2*c12y3*c13.x - c11x3*c20.y*c13y2 - c12x3*c20y2*c13.y - c10.x*c11x2*c11.y*c13y2 +
+            c10.y*c11.x*c11y2*c13x2 - 3*c10.x*c10y2*c13x2*c13.y - c10.x*c11y2*c12x2*c13.y + c10.y*c11x2*c12y2*c13.x -
+            c11.x*c11y2*c20.y*c13x2 + 3*c10x2*c10.y*c13.x*c13y2 + c10x2*c11.x*c12.y*c13y2 +
+            2*c10x2*c11.y*c12.x*c13y2 - 2*c10y2*c11.x*c12.y*c13x2 - c10y2*c11.y*c12.x*c13x2 + c11x2*c20.x*c11.y*c13y2 -
+            3*c10.x*c20y2*c13x2*c13.y + 3*c10.y*c20x2*c13.x*c13y2 + c11.x*c20x2*c12.y*c13y2 - 2*c11.x*c20y2*c12.y*c13x2 +
+            c20.x*c11y2*c12x2*c13.y - c11.y*c12.x*c20y2*c13x2 - c10x2*c12.x*c12y2*c13.y - 3*c10x2*c20.y*c13.x*c13y2 +
+            3*c10y2*c20.x*c13x2*c13.y + c10y2*c12x2*c12.y*c13.x - c11x2*c20.y*c12y2*c13.x + 2*c20x2*c11.y*c12.x*c13y2 +
+            3*c20.x*c20y2*c13x2*c13.y - c20x2*c12.x*c12y2*c13.y - 3*c20x2*c20.y*c13.x*c13y2 + c12x2*c20y2*c12.y*c13.x
+    );
+    var roots = poly.getRootsInInterval(0,1);
+    module.exports.Utils.removeMultipleRootsIn01(roots);
+
+    for ( var i = 0; i < roots.length; i++ ) {
+        var s = roots[i];
+        var xRoots = new Polynomial(
+            c13.x,
+            c12.x,
+            c11.x,
+            c10.x - c20.x - s*c21.x - s*s*c22.x - s*s*s*c23.x
+        ).getRoots();
+        var yRoots = new Polynomial(
+            c13.y,
+            c12.y,
+            c11.y,
+            c10.y - c20.y - s*c21.y - s*s*c22.y - s*s*s*c23.y
+        ).getRoots();
+
+        if ( xRoots.length > 0 && yRoots.length > 0 ) {
+            var TOLERANCE = 1e-4;
+
+            checkRoots:
+                for ( var j = 0; j < xRoots.length; j++ ) {
+                    var xRoot = xRoots[j];
+
+                    if ( 0 <= xRoot && xRoot <= 1 ) {
+                        for ( var k = 0; k < yRoots.length; k++ ) {
+                            if ( Math.abs( xRoot - yRoots[k] ) < TOLERANCE ) {
+                                var v = c23.multiply(s * s * s).add(c22.multiply(s * s).add(c21.multiply(s).add(c20)));
+                                result.points.push(new Point2D(v.x, v.y));
+                                break checkRoots;
+                            }
+                        }
+                    }
+                }
+        }
+    }
+
+    return result;
+};
+
+/**
+ *  intersectBezier3Ellipse
+ *
+ *  @param {Point2D} p1
+ *  @param {Point2D} p2
+ *  @param {Point2D} p3
+ *  @param {Point2D} p4
+ *  @param {Point2D} ec
+ *  @param {Number} rx
+ *  @param {Number} ry
+ *  @returns {Intersection}
+ */
+module.exports.intersectBezier3Ellipse = function(p1, p2, p3, p4, ec, rx, ry) {
+    var a, b, c, d;       // temporary variables
+    var c3, c2, c1, c0;   // coefficients of cubic
+    var result = new Intersection();
+
+    // Calculate the coefficients of cubic polynomial
+    a = p1.multiply(-1);
+    b = p2.multiply(3);
+    c = p3.multiply(-3);
+    d = a.add(b.add(c.add(p4)));
+    c3 = new Vector2D(d.x, d.y);
+
+    a = p1.multiply(3);
+    b = p2.multiply(-6);
+    c = p3.multiply(3);
+    d = a.add(b.add(c));
+    c2 = new Vector2D(d.x, d.y);
+
+    a = p1.multiply(-3);
+    b = p2.multiply(3);
+    c = a.add(b);
+    c1 = new Vector2D(c.x, c.y);
+
+    c0 = new Vector2D(p1.x, p1.y);
+
+    var rxrx  = rx*rx;
+    var ryry  = ry*ry;
+    var poly = new Polynomial(
+        c3.x*c3.x*ryry + c3.y*c3.y*rxrx,
+        2*(c3.x*c2.x*ryry + c3.y*c2.y*rxrx),
+        2*(c3.x*c1.x*ryry + c3.y*c1.y*rxrx) + c2.x*c2.x*ryry + c2.y*c2.y*rxrx,
+        2*c3.x*ryry*(c0.x - ec.x) + 2*c3.y*rxrx*(c0.y - ec.y) +
+            2*(c2.x*c1.x*ryry + c2.y*c1.y*rxrx),
+        2*c2.x*ryry*(c0.x - ec.x) + 2*c2.y*rxrx*(c0.y - ec.y) +
+            c1.x*c1.x*ryry + c1.y*c1.y*rxrx,
+        2*c1.x*ryry*(c0.x - ec.x) + 2*c1.y*rxrx*(c0.y - ec.y),
+        c0.x*c0.x*ryry - 2*c0.y*ec.y*rxrx - 2*c0.x*ec.x*ryry +
+            c0.y*c0.y*rxrx + ec.x*ec.x*ryry + ec.y*ec.y*rxrx - rxrx*ryry
+    );
+    var roots = poly.getRootsInInterval(0,1);
+    module.exports.Utils.removeMultipleRootsIn01(roots);
+
+    for ( var i = 0; i < roots.length; i++ ) {
+        var t = roots[i];
+        var v = c3.multiply(t * t * t).add(c2.multiply(t * t).add(c1.multiply(t).add(c0)));
+        result.points.push(new Point2D(v.x, v.y));
+    }
+
+    return result;
+};
+
+
+/**
+ *  intersectBezier3Line
+ *
+ *  Many thanks to Dan Sunday at SoftSurfer.com.  He gave me a very thorough
+ *  sketch of the algorithm used here.  Without his help, I'm not sure when I
+ *  would have figured out this intersection problem.
+ *
+ *  @param {Point2D} p1
+ *  @param {Point2D} p2
+ *  @param {Point2D} p3
+ *  @param {Point2D} p4
+ *  @param {Point2D} a1
+ *  @param {Point2D} a2
+ *  @returns {Intersection}
+ */
+module.exports.intersectBezier3Line = function(p1, p2, p3, p4, a1, a2) {
+    var a, b, c, d;       // temporary variables
+    var c3, c2, c1, c0;   // coefficients of cubic
+    var cl;               // c coefficient for normal form of line
+    var n;                // normal for normal form of line
+    var min = a1.min(a2); // used to determine if point is on line segment
+    var max = a1.max(a2); // used to determine if point is on line segment
+    var result = new Intersection();
+
+    // Start with Bezier using Bernstein polynomials for weighting functions:
+    //     (1-t^3)P1 + 3t(1-t)^2P2 + 3t^2(1-t)P3 + t^3P4
+    //
+    // Expand and collect terms to form linear combinations of original Bezier
+    // controls.  This ends up with a vector cubic in t:
+    //     (-P1+3P2-3P3+P4)t^3 + (3P1-6P2+3P3)t^2 + (-3P1+3P2)t + P1
+    //             /\                  /\                /\       /\
+    //             ||                  ||                ||       ||
+    //             c3                  c2                c1       c0
+
+    // Calculate the coefficients
+    a = p1.multiply(-1);
+    b = p2.multiply(3);
+    c = p3.multiply(-3);
+    d = a.add(b.add(c.add(p4)));
+    c3 = new Vector2D(d.x, d.y);
+
+    a = p1.multiply(3);
+    b = p2.multiply(-6);
+    c = p3.multiply(3);
+    d = a.add(b.add(c));
+    c2 = new Vector2D(d.x, d.y);
+
+    a = p1.multiply(-3);
+    b = p2.multiply(3);
+    c = a.add(b);
+    c1 = new Vector2D(c.x, c.y);
+
+    c0 = new Vector2D(p1.x, p1.y);
+
+    // Convert line to normal form: ax + by + c = 0
+    // Find normal to line: negative inverse of original line's slope
+    n = new Vector2D(a1.y - a2.y, a2.x - a1.x);
+
+    // Determine new c coefficient
+    cl = a1.x*a2.y - a2.x*a1.y;
+
+    // ?Rotate each cubic coefficient using line for new coordinate system?
+    // Find roots of rotated cubic
+    roots = new Polynomial(
+        n.dot(c3),
+        n.dot(c2),
+        n.dot(c1),
+        n.dot(c0) + cl
+    ).getRoots();
+
+    // Any roots in closed interval [0,1] are intersections on Bezier, but
+    // might not be on the line segment.
+    // Find intersections and calculate point coordinates
+    for ( var i = 0; i < roots.length; i++ ) {
+        var t = roots[i];
+
+        if ( 0 <= t && t <= 1 ) {
+            // We're within the Bezier curve
+            // Find point on Bezier
+            var p5 = p1.lerp(p2, t);
+            var p6 = p2.lerp(p3, t);
+            var p7 = p3.lerp(p4, t);
+
+            var p8 = p5.lerp(p6, t);
+            var p9 = p6.lerp(p7, t);
+
+            var p10 = p8.lerp(p9, t);
+
+            // See if point is on line segment
+            // Had to make special cases for vertical and horizontal lines due
+            // to slight errors in calculation of p10
+            if ( a1.x == a2.x ) {
+                if ( min.y <= p10.y && p10.y <= max.y ) {
+                    result.appendPoint( p10 );
+                }
+            } else if ( a1.y == a2.y ) {
+                if ( min.x <= p10.x && p10.x <= max.x ) {
+                    result.appendPoint( p10 );
+                }
+            } else if (min.x <= p10.x && p10.x <= max.x && min.y <= p10.y && p10.y <= max.y) {
+                result.appendPoint( p10 );
+            }
+        }
+    }
+
+    return result;
+};
+
